LayoutTests/fast/css/font-systemFontID-parsing-expected.txt

Tests 'font' CSS property parsing with system font values.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS testDivInvalid.style.getPropertyCSSValue('font') is null
FAIL testDivValid.style.getPropertyCSSValue('font') should be caption (of type string). Was null (of type object).
PASS testDivInvalid.style.getPropertyCSSValue('font-family') is null
PASS testDivValid.style.getPropertyCSSValue('font-family') is null
PASS successfullyParsed is true

TEST COMPLETE
